notifier_service/requirements.txt
flask
flask-socketio
eventlet
pika
python-dotenv
flask_cors
kombu